Refactor algorithms test module in extras
This removes a number of redundant germ selection tests and moves the rest of the (now much smaller faster) tests alongside the rest in the 'algorithms' folder. Deletes 'algorithmsb' directory.
