+++ test/test_packages/algorithms/test_germselection.py
@@ -1,0 +1,54 @@
+import pygsti
+import numpy as _np
+from pygsti.circuits import Circuit
+from pygsti.baseobjs import Label
+from pygsti.modelpacks import smq1Q_XY as std
+from ..algorithms.algorithmsTestCase import AlgorithmTestCase
+
+
+class GermSelectionTestCase(AlgorithmTestCase):
+
+    #test with worst score_func
+    def test_germsel_greedy(self):
+        threshold             = 1e6
+        randomizationStrength = 1e-3
+        neighborhoodSize      = 2
+        gatesetNeighborhood   = pygsti.alg.randomize_model_list([std.target_model()],
+                                                                randomization_strength=randomizationStrength,
+                                                                num_copies=neighborhoodSize, seed=2014)
+
+        max_length   = 4
+        gates        = std.target_model().operations.keys()
+        superGermSet = pygsti.circuits.list_all_circuits_without_powers_and_cycles(gates, max_length)
+
+        pygsti.alg.find_germs_breadthfirst(gatesetNeighborhood, superGermSet,
+                                           randomize=False, seed=2014, score_func='worst',
+                                           threshold=threshold, verbosity=1, op_penalty=1.0,
+                                           mem_limit=2*1024000)
+                                           
+    def test_germsel_driver_greedy(self):
+        #GREEDY
+        options = {'threshold': 1e6 }
+        germs = pygsti.alg.find_germs(std.target_model(), randomize=True, randomization_strength=1e-3,
+                                      num_gs_copies=2, seed=2017, candidate_germ_counts={3: 'all upto', 4: 10, 5:10, 6:10},
+                                      candidate_seed=2017, force="singletons", algorithm='greedy',
+                                      algorithm_kwargs=options, mem_limit=None, comm=None,
+                                      profiler=None, verbosity=1)
+          
+    def test_germsel_driver_grasp(self):
+        #more args
+        options = {'threshold': 1e6 , 'return_all': True}
+        germs2 = pygsti.alg.find_germs(std.target_model(), randomize=True, randomization_strength=1e-3,
+                                       num_gs_copies=2, seed=2017, candidate_germ_counts={3: 'all upto', 4: 10, 5:10, 6:10},
+                                       candidate_seed=2017, force="singletons", algorithm='grasp',
+                                       algorithm_kwargs=options, mem_limit=None,
+                                       profiler=None, verbosity=1)
+
+    def test_germsel_driver_slack(self):
+        #SLACK
+        options = dict(fixed_slack=False, slack_frac=0.1)
+        germs = pygsti.alg.find_germs(std.target_model(), randomize=True, randomization_strength=1e-3,
+                                      num_gs_copies=2, seed=2017, candidate_germ_counts={3: 'all upto', 4: 10, 5:10, 6:10},
+                                      candidate_seed=2017, force="singletons", algorithm='slack',
+                                      algorithm_kwargs=options, mem_limit=None, comm=None,
+                                      profiler=None, verbosity=1)
